cli/requirements.txt
click>=8.0.0
httpx>=0.24.0
GitPython>=3.1.0
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
pydantic>=2.0.0
loguru>=0.7.0

